feat: refactor banner & add navbar functionality
